requirements.txt
psutil>=5.9.0
netifaces>=0.11.0
py-cpuinfo>=9.0.0
rich>=13.0.0